move bench logic into internal package
so it somehow makes Goland IDE able to compile
correctly the module.

It's unfortunate that somehow Goland trips with Go
code that lives in the `main` package, but thought
it may be acceptable to the mantainer/s to move logic
to an internal package as the pattern is already used
in this repo and is effectively a noop.
